Ensure fixup doesn't get confused when already done manually
Issue #6067

The issue was that we were not checking that the inverse had not already been fixed up before changing it. Added a check to stop this. Also found related issue that if shared PK/FK of Added entity was set manually, then it shouldn't be overwritten with same value by fixup since this will stop it being processed correctly later by DetectChanges.